+++ test/fixtures/apps/app-die-ignore-code/app/router.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = app => {
+  app.get('/uncaughtException', function*() {
+    setTimeout(() => {
+      const error = new Error('MockError');
+      error.code = 'EMOCKERROR';
+      throw error;
+    }, 100);
+  });
+};