+++ test/fixtures/apps/app-die-ignore-code/config/config.default.js
@@ -1,0 +1,3 @@
+exports.keys = 'foo';
+
+exports.serverGracefulIgnoreCode = ['EMOCKERROR'];