+++ test/fixtures/apps/app-ts/config/config.ts
@@ -14,5 +14,12 @@ export default () => {
     },
   };
 
+  config.httpclient = {
+    useHttpClientNext: false,
+    request: {
+      timing: true,
+    },
+  };
+
   return config;
 }