+++ test/fixtures/apps/httpclient-next-overwrite/config/config.default.js
@@ -1,0 +1,8 @@
+'use strict';
+
+exports.httpclient = {
+  useHttpClientNext: true,
+  request: {
+    timeout: 99,
+  },
+};