+++ test/lib/cluster/app_worker.test.js
@@ -1,10 +1,7 @@
-'use strict';
-
 const net = require('net');
 const request = require('supertest');
 const address = require('address');
 const assert = require('assert-extends');
-const sleep = require('mz-modules/sleep');
 const utils = require('../../utils');
 
 const DEFAULT_BAD_REQUEST_HTML = `<html>
@@ -104,7 +101,7 @@ describe('test/lib/cluster/app_worker.test.js', () => {
         version[0] > 9) {
         html = new RegExp(
           'GET /foo bar HTTP/1.1\r\nHost: 127.0.0.1:\\d+\r\nAccept-Encoding: gzip, ' +
-          'deflate\r\nUser-Agent: node-superagent/\\d+\\.\\d+\\.\\d+\r\nConnection: close\r\n\r\n');
+          'deflate\r\nConnection: close\r\n\r\n');
       }
 
       // customized client error response
@@ -112,7 +109,7 @@ describe('test/lib/cluster/app_worker.test.js', () => {
       test1.request().path = '/foo bar';
       await test1.expect(html)
         .expect('foo', 'bar')
-        .expect('content-length', '134')
+        .expect('content-length', '99')
         .expect(418);
 
       // customized client error handle function throws
@@ -123,7 +120,7 @@ describe('test/lib/cluster/app_worker.test.js', () => {
 
     it('should not log when there is no rawPacket', async () => {
       await connect(app.port);
-      await sleep(1000);
+      await utils.sleep(1000);
       app.expect('stderr', /HPE_INVALID_EOF_STATE/);
       app.notExpect('stderr', /A client/);
     });